E2-1628 refactor request class part 1 (#276)
* add LPRequesting interface

* add LPRequest, LPRequesting, RequestFactory

* add apiconfig

* minor cleanup

* rename old Request class to RequestOld, add new Request class

* add some methods, change some to private

* add countaggregator and featureflag to recFactory

* minor change to getInstance in reqFactory
