+++ AndroidSDKCore/src/main/java/com/leanplum/internal/APIConfig.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2018, Leanplum, Inc. All rights reserved.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.leanplum.internal;
+
+public class APIConfig {
+  private final FeatureFlagManager featureFlagManager;
+  private final CountAggregator countAggregator;
+  private String appId;
+  private String accessKey;
+  private String token;
+
+  public APIConfig(FeatureFlagManager featureFlagManager, CountAggregator countAggregator) {
+    this.featureFlagManager = featureFlagManager;
+    this.countAggregator = countAggregator;
+  }
+
+  public void setAppId(String appId, String accessKey) {
+    this.appId = appId;
+    this.accessKey = accessKey;
+
+    this.countAggregator.incrementCount("set_app_id");
+  }
+
+  public void loadToken(String token) {
+    this.token = token;
+  }
+}
+
