+++ AndroidSDKCore/src/main/java/com/leanplum/internal/Log.java
@@ -188,7 +188,7 @@ public class Log {
       HashMap<String, Object> params = new HashMap<>();
       params.put(Constants.Params.TYPE, Constants.Values.SDK_LOG);
       params.put(Constants.Params.MESSAGE, message);
-      Request.post(Constants.Methods.LOG, params).sendEventually();
+      RequestOld.post(Constants.Methods.LOG, params).sendEventually();
     } catch (Throwable t) {
       android.util.Log.e("Leanplum", "Unable to send log.", t);
     } finally {
