+++ AndroidSDKCore/src/main/java/com/leanplum/internal/RequestSequenceRecorder.java
@@ -2,15 +2,15 @@ package com.leanplum.internal;
 
 /** Records request call sequence of read/write operations to database. */
 public interface RequestSequenceRecorder {
-  /** Executes before database read in Request. */
+  /** Executes before database read in RequestOld. */
   void beforeRead();
 
-  /** Executes after database read in Request. */
+  /** Executes after database read in RequestOld. */
   void afterRead();
 
-  /** Executes before database write in Request. */
+  /** Executes before database write in RequestOld. */
   void beforeWrite();
 
-  /** Executes after database write in Request. */
+  /** Executes after database write in RequestOld. */
   void afterWrite();
 }
