+++ AndroidSDKCore/src/main/java/com/leanplum/internal/Requesting.java
@@ -1,0 +1,14 @@
+package com.leanplum.internal;
+
+import org.json.JSONObject;
+
+public interface Requesting {
+
+  interface ResponseCallback {
+    void response(JSONObject response);
+  }
+
+  interface ErrorCallback {
+    void error(Exception e);
+  }
+}
