+++ AndroidSDKCore/src/main/java/com/leanplum/internal/Util.java
@@ -454,7 +454,7 @@ public class Util {
     Uri.Builder builder = new Uri.Builder();
     for (Map.Entry<String, Object> pair : params.entrySet()) {
       if (pair.getValue() == null) {
-        Log.w("Request parameter for key: " + pair.getKey() + " is null.");
+        Log.w("RequestOld parameter for key: " + pair.getKey() + " is null.");
         continue;
       }
       builder.appendQueryParameter(pair.getKey(), pair.getValue().toString());
@@ -554,7 +554,7 @@ public class Util {
     urlConnection.setInstanceFollowRedirects(true);
     Context context = Leanplum.getContext();
     urlConnection.setRequestProperty("User-Agent",
-        getApplicationName(context) + "/" + getVersionName() + "/" + Request.appId() + "/" +
+        getApplicationName(context) + "/" + getVersionName() + "/" + RequestOld.appId() + "/" +
             Constants.CLIENT + "/" + Constants.LEANPLUM_VERSION + "/" + getSystemName() + "/" +
             getSystemVersion() + "/" + Constants.LEANPLUM_PACKAGE_IDENTIFIER);
     return urlConnection;
@@ -893,7 +893,7 @@ public class Util {
       params.put("stackTrace", stringWriter.toString());
 
       params.put(Params.VERSION_NAME, versionName);
-      Request.post(Methods.LOG, params).send();
+      RequestOld.post(Methods.LOG, params).send();
     } catch (Throwable t2) {
       Log.e("Unable to send error report.", t2);
     }
