+++ AndroidSDKCore/src/main/java/com/leanplum/internal/VarCache.java
@@ -341,7 +341,7 @@ public class VarCache {
     }
     Context context = Leanplum.getContext();
     SharedPreferences defaults = context.getSharedPreferences(LEANPLUM, Context.MODE_PRIVATE);
-    if (Request.token() == null) {
+    if (RequestOld.token() == null) {
       applyVariableDiffs(
           new HashMap<String, Object>(),
           new HashMap<String, Object>(),
@@ -354,7 +354,7 @@ public class VarCache {
     }
     try {
       // Crypt functions return input text if there was a problem.
-      AESCrypt aesContext = new AESCrypt(Request.appId(), Request.token());
+      AESCrypt aesContext = new AESCrypt(RequestOld.appId(), RequestOld.token());
       String variables = aesContext.decodePreference(
           defaults, Constants.Defaults.VARIABLES_KEY, "{}");
       String messages = aesContext.decodePreference(
@@ -377,7 +377,7 @@ public class VarCache {
       String deviceId = aesContext.decodePreference(defaults, Constants.Params.DEVICE_ID, null);
       if (deviceId != null) {
         if (Util.isValidDeviceId(deviceId)) {
-          Request.setDeviceId(deviceId);
+          RequestOld.setDeviceId(deviceId);
         } else {
           Log.w("Invalid stored device id found: \"" + deviceId + "\"; discarding.");
         }
@@ -385,7 +385,7 @@ public class VarCache {
       String userId = aesContext.decodePreference(defaults, Constants.Params.USER_ID, null);
       if (userId != null) {
         if (Util.isValidUserId(userId)) {
-          Request.setUserId(userId);
+          RequestOld.setUserId(userId);
         } else {
           Log.w("Invalid stored user id found: \"" + userId + "\"; discarding.");
         }
@@ -406,7 +406,7 @@ public class VarCache {
     if (Constants.isNoop()) {
       return;
     }
-    if (Request.token() == null) {
+    if (RequestOld.token() == null) {
       return;
     }
     Context context = Leanplum.getContext();
@@ -414,7 +414,7 @@ public class VarCache {
     SharedPreferences.Editor editor = defaults.edit();
 
     // Crypt functions return input text if there was a problem.
-    AESCrypt aesContext = new AESCrypt(Request.appId(), Request.token());
+    AESCrypt aesContext = new AESCrypt(RequestOld.appId(), RequestOld.token());
     String variablesCipher = aesContext.encrypt(JsonConverter.toJson(diffs));
     editor.putString(Constants.Defaults.VARIABLES_KEY, variablesCipher);
 
@@ -459,8 +459,8 @@ public class VarCache {
           aesContext.encrypt(JsonConverter.toJson(variantDebugInfo)));
     }
 
-    editor.putString(Constants.Params.DEVICE_ID, aesContext.encrypt(Request.deviceId()));
-    editor.putString(Constants.Params.USER_ID, aesContext.encrypt(Request.userId()));
+    editor.putString(Constants.Params.DEVICE_ID, aesContext.encrypt(RequestOld.deviceId()));
+    editor.putString(Constants.Params.USER_ID, aesContext.encrypt(RequestOld.userId()));
     editor.putString(Constants.Keys.LOGGING_ENABLED,
         aesContext.encrypt(String.valueOf(Constants.loggingEnabled)));
     SharedPreferencesUtil.commitChanges(editor);
@@ -706,7 +706,7 @@ public class VarCache {
         params.put(Constants.Params.ACTION_DEFINITIONS, JsonConverter.toJson(actionDefinitions));
       }
       params.put(Constants.Params.FILE_ATTRIBUTES, JsonConverter.toJson(fileAttributes));
-      Request.post(Constants.Methods.SET_VARS, params).sendIfConnected();
+      RequestOld.post(Constants.Methods.SET_VARS, params).sendIfConnected();
     }
 
     return changed;
@@ -751,7 +751,7 @@ public class VarCache {
           Map<String, Object> params = new HashMap<>();
           params.put(Constants.Params.DATA, fileData.toString());
 
-          Request.post(Constants.Methods.UPLOAD_FILE, params).sendFilesNow(filenames,
+          RequestOld.post(Constants.Methods.UPLOAD_FILE, params).sendFilesNow(filenames,
               streams);
 
           filenames = new ArrayList<>();
@@ -782,7 +782,7 @@ public class VarCache {
     if (filenames.size() > 0) {
       Map<String, Object> params = new HashMap<>();
       params.put(Constants.Params.DATA, fileData.toString());
-      Request.post(Constants.Methods.UPLOAD_FILE, params).sendFilesNow(filenames, streams);
+      RequestOld.post(Constants.Methods.UPLOAD_FILE, params).sendFilesNow(filenames, streams);
     }
   }
 
@@ -882,7 +882,7 @@ public class VarCache {
     if (userAttributes == null) {
       Context context = Leanplum.getContext();
       SharedPreferences defaults = context.getSharedPreferences(LEANPLUM, Context.MODE_PRIVATE);
-      AESCrypt aesContext = new AESCrypt(Request.appId(), Request.token());
+      AESCrypt aesContext = new AESCrypt(RequestOld.appId(), RequestOld.token());
       try {
         userAttributes = JsonConverter.fromJson(
             aesContext.decodePreference(defaults, Constants.Defaults.ATTRIBUTES_KEY, "{}"));
@@ -895,7 +895,7 @@ public class VarCache {
   }
 
   public static void saveUserAttributes() {
-    if (Constants.isNoop() || Request.appId() == null || userAttributes == null) {
+    if (Constants.isNoop() || RequestOld.appId() == null || userAttributes == null) {
       return;
     }
     Context context = Leanplum.getContext();
@@ -903,7 +903,7 @@ public class VarCache {
     SharedPreferences.Editor editor = defaults.edit();
     // Crypt functions return input text if there was a problem.
     String plaintext = JsonConverter.toJson(userAttributes);
-    AESCrypt aesContext = new AESCrypt(Request.appId(), Request.token());
+    AESCrypt aesContext = new AESCrypt(RequestOld.appId(), RequestOld.token());
     editor.putString(Constants.Defaults.ATTRIBUTES_KEY, aesContext.encrypt(plaintext));
     SharedPreferencesUtil.commitChanges(editor);
