+++ AndroidSDKLocation/src/main/java/com/leanplum/LocationManagerImplementation.java
@@ -447,7 +447,7 @@ class LocationManagerImplementation implements
   }
 
   /**
-   * Request location for user location update if googleApiClient is connected.
+   * RequestOld location for user location update if googleApiClient is connected.
    */
   // Suppressing missing permission warning which since it is up to client to add location
   // permission to their manifest.
